test: mock.js 파일 분리
